pyproject.toml
[project]
name = "mcp-demos"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agents>=1.4.0",
    "fastmcp>=2.3.3",
    "httpx>=0.28.1",
    "langfuse>=2.60.7",
    "langtrace-python-sdk>=3.8.18",
    "mcp>=1.8.1",
    "openai>=1.78.1",
    "openai-agents>=0.0.14",
    "pandas>=2.2.3",
    "pyfiglet>=1.0.2",
    "python-dotenv>=1.1.0",
    "tensorflow-io-gcs-filesystem==0.31.0",
    "playwright>=1.52.0",
    "markdownify>=1.1.0",
    "beautifulsoup4>=4.13.4",
    "html2text>=2025.4.15",
    "lxml>=5.4.0",
]
